The 'first_time' variable has been removed: it is not necessary to determine the candidate images to be swapped.
The 'if (images.length > 0)' conditional has been removed as well: it is
not necessary because if there is no candidate image to be swapped, the
'swappable_images' variable will be empty and the each loop will not do
anything with it.

All the variables (with the exception of $win and th) have been declared
within the Flojo() function (local variables). Local variables are deleted
when the function is completed.

Finally, it has been added/modified some comments to make the script more
understandable.